add verbose print on errors that can hide starknet data
Currently, there is some data missing to be printed out
by starknet-rs. ContractError reason for instance is never printed.
The idea here is to have a bit more context with sozo if a user
is stuck during migration.
This commit starts work for the #1409 issue to add more logging
to sozo.