refactor(katana-db): convert `StorageChangeSet` into a normal table (#1774)
Resolves #1762 

Convert `StorageChangeSet` into a normal table instead of  dupsort.

This change is so to allow the block list of the `StorageChangeSet` to grow without being bounded by the value size of a `DUPSORT` table. 

The max size for a DUPSORT table value is equal to the max size of the table key. So, when the list cardinality increases too much, the resulting subkey size will become larger than the max size. This will result in a `MDBX_BAD_VALSIZE` error.