fix(core): upgrade checks for key members (#3285)
In the way we check if a model can be upgraded, some cases for `key` members are not well managed.

Here are the `key` member upgrade rules that are checked according to the `key` member type:
- `Primitive`: same than for a `value` member, the new type must have the same or a bigger size than the old type (`u8` -> `u16`),
- `Struct`: new members are NOT allowed. Existing member types must follow the `key` member upgrade rules,
- `Enum`: new variants are allowed. Existing variant data types must follow the `key` member upgrade rules,
- `Tuple`: tuple items must follow the `key` member upgrade rules and the tuple size must remain the same,
- `Fixed-Size Array`: the array item type must follow the `key` member upgrade rules and the array size must remain the same,
- `Array`: the array item type must follow the `key` member upgrade rules,
- `ByteArray`: always allowed.

## Model reading with enum/option key

Since the introduction of `DojoStore`, when a model is read, keys are serialized using `Serde` but read values were serialized with `DojoStore`. That means, when we try to deserialize the whole data keys+values using `DojoStore` it fails due to the fact that `Serde` and `DojoStore` don't handle variant indexes the same way.

So, this commit updates the `ModelParser::deserialize()` function to:
- still use `Serde` to deserialize events and legacy models,
- deserialize field by field, using `Serde` for key fields and `DojoStore` for value fields.

---------

Co-authored-by: Ammar Arif <evergreenkary@gmail.com>